requirements.txt
alembic==0.8.7
arrow==0.8.0
beautifulsoup4==4.5.0
click==6.6
cookies==2.2.1
Flask==0.11.1
Flask-Login==0.3.2
Flask-Migrate==1.8.1
Flask-Script==2.0.5
Flask-SQLAlchemy==2.1
funcsigs==1.0.2
itsdangerous==0.24
Jinja2==2.8
Mako==1.0.4
MarkupSafe==0.23
mock==2.0.0
pbr==1.10.0
python-dateutil==1.5
python-editor==1.0.1
python-forecastio==1.3.4
pytz==2016.6.1
requests==2.10.0
responses==0.5.1
six==1.10.0
SQLAlchemy==1.0.14
stravalib==0.5.0
units==0.6
Werkzeug==0.11.10
